Add dialog box check on event deletion

+++ Permastead/ViewModels/Views/EventsViewModel.cs
@@ -1,11 +1,16 @@
 using System;
 using System.Collections.ObjectModel;
 using System.Linq;
+using System.Threading.Tasks;
+using System.Windows.Input;
+using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.Models.TreeDataGrid;
 using Avalonia.Media;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+
+using Ursa.Controls;
 
 using Models;
 using Permastead.ViewModels.Dialogs;
@@ -38,6 +43,28 @@ public partial class  EventsViewModel : ViewModelBase
     private AnEvent _currentItem;
     
     public FlatTreeDataGridSource<AnEvent> EventsSource { get; set; }
+    
+    private readonly string _shortMessage = "Are you sure you want to delete this event?";
+    private string _message;
+    private string? _title = "Deletion Confirmation";
+    
+    public ObservableCollection<MessageBoxIcon> Icons { get; set; }
+    
+    private MessageBoxIcon _selectedIcon;
+    public MessageBoxIcon SelectedIcon
+    {
+        get => _selectedIcon;
+        set => SetProperty(ref _selectedIcon, value);
+    }
+
+    private MessageBoxResult _result;
+    public MessageBoxResult Result
+    {
+        get => _result;
+        set => SetProperty(ref _result, value);
+    }
+    
+    public ICommand YesNoCommand { get; set; }
 
     [RelayCommand]
     private void RefreshData()
@@ -69,11 +96,23 @@ public partial class  EventsViewModel : ViewModelBase
                 CurrentItem = new AnEvent();
             }
             
+            YesNoCommand = new AsyncRelayCommand(OnYesNoAsync);
+            
+            Icons = new ObservableCollection<MessageBoxIcon>(
+                Enum.GetValues<MessageBoxIcon>());
+            SelectedIcon = MessageBoxIcon.Question;
+            _message = _shortMessage;
+            
         }
         catch (Exception)
         {
             throw;
         }
+    }
+    
+    private async Task OnYesNoAsync()
+    {
+        await Show(MessageBoxButton.YesNo);
     }
     
     public void RefreshEvents()
@@ -223,15 +262,26 @@ public partial class  EventsViewModel : ViewModelBase
     }
     
     [RelayCommand]
-    private void RemoveEvent()
+    private async void RemoveEvent()
     {
         if (CurrentItem != null)
         {
             var anEvent = CurrentItem;
+
+            await OnYesNoAsync();
+
+            if (Result == MessageBoxResult.Yes)
+            {
+                //remove the record
+                EventsService.DeleteEvent(AppSession.ServiceMode, CurrentItem);
+                RefreshEvents();
+            }
             
-            //remove the record
-            EventsService.DeleteEvent(AppSession.ServiceMode, CurrentItem);
-            RefreshEvents();
         }
+    }
+    
+    private async Task Show(MessageBoxButton button)
+    {
+        Result = await MessageBox.ShowAsync(_message, _title, icon: SelectedIcon, button:button);
     }
 }
